through sub_80E9E44

+++ include/menu_indicators.h
@@ -6,7 +6,7 @@
 // Exported RAM declarations
 
 // Exported ROM declarations
-u8 AddScrollIndicatorArrowPairParametrized(u8, u8, u8, u8, s32, u8, u8, u16 *);
+u8 AddScrollIndicatorArrowPairParametrized(u8, u8, u8, u8, s32, u16, u16, u16 *);
 void RemoveScrollIndicatorArrowPair(u8);
 
 #endif //GUARD_MENU_INDICATORS_H
